+++ src/com/upokecenter/util/FastInteger.java
@@ -28,7 +28,7 @@ at: http://upokecenter.com/d/
       public static MutableNumber FromBigInteger(BigInteger bigintVal) {
         MutableNumber mnum = new MutableNumber(0);
         if (bigintVal.signum() < 0) {
- throw new IllegalArgumentException("bigintVal's sign (" + Long.toString((long)bigintVal.signum()) + ") is less than " + "0");
+ throw new IllegalArgumentException("bigintVal's sign (" + Integer.toString((int)bigintVal.signum()) + ") is less than " + "0");
 }
         byte[] bytes = bigintVal.toByteArray(true);
         int len = bytes.length;
@@ -61,7 +61,7 @@ at: http://upokecenter.com/d/
 
       public MutableNumber (int val) {
         if (val < 0) {
- throw new IllegalArgumentException("val (" + Long.toString((long)val) + ") is less than " + "0");
+ throw new IllegalArgumentException("val (" + Integer.toString((int)val) + ") is less than " + "0");
 }
         this.data = new int[4];
         this.wordCount = (val == 0) ? 0 : 1;
@@ -75,7 +75,7 @@ at: http://upokecenter.com/d/
      */
       public MutableNumber SetInt(int val) {
         if (val < 0) {
- throw new IllegalArgumentException("val (" + Long.toString((long)val) + ") is less than " + "0");
+ throw new IllegalArgumentException("val (" + Integer.toString((int)val) + ") is less than " + "0");
 }
         this.wordCount = (val == 0) ? 0 : 1;
         this.data[0] = ((int)(val & 0xFFFFFFFFL));
@@ -144,7 +144,7 @@ at: http://upokecenter.com/d/
      */
       public MutableNumber Multiply(int multiplicand) {
         if (multiplicand < 0) {
- throw new IllegalArgumentException("multiplicand (" + Long.toString((long)multiplicand) + ") is less than " + "0");
+ throw new IllegalArgumentException("multiplicand (" + Integer.toString((int)multiplicand) + ") is less than " + "0");
 } else if (multiplicand != 0) {
           int carry = 0;
           if (this.wordCount == 0) {
@@ -288,7 +288,7 @@ at: http://upokecenter.com/d/
       public MutableNumber SubtractInt(
         int other) {
         if (other < 0) {
- throw new IllegalArgumentException("other (" + Long.toString((long)other) + ") is less than " + "0");
+ throw new IllegalArgumentException("other (" + Integer.toString((int)other) + ") is less than " + "0");
 } else if (other != 0) {
           {
             // Ensure a length of at least 1
@@ -402,7 +402,7 @@ at: http://upokecenter.com/d/
      */
       public MutableNumber Add(int augend) {
         if (augend < 0) {
- throw new IllegalArgumentException("augend (" + Long.toString((long)augend) + ") is less than " + "0");
+ throw new IllegalArgumentException("augend (" + Integer.toString((int)augend) + ") is less than " + "0");
 } else if (augend != 0) {
           int carry = 0;
           // Ensure a length of at least 1
